# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Exam project API",
    "description": "API Documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token obtained after login"
    }
  },
  "paths": {
...
    "/brands/": {
      "get": {
        "description": "Getting all the brands",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "adding new brand",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/brand"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/brands/{id}": {
      "put": {
        "description": "editing/changing a brand",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "delete/remove a brand",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/admin/brands": {
      "get": {
        "description": "Get brands and render be display",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "brand": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "swagger brand"
        }
      },
      "required": [
        "name"
      ]
    },
...
  }
}